requirements.txt
astroid==1.4.4
beautifulsoup4==4.5.3
blinker==1.4
click==6.7
configparser==3.5.0
Flask==0.12.1
Flask-Compress==1.4.0
Flask-DebugToolbar==0.10.1
Flask-RBAC==0.3.0
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.2.1
logilab-common==1.1.0
MarkupSafe==1.0
numpy==1.12.1
olefile==0.44
pandas==0.20.1
Pillow==4.1.1
py==1.4.33
PyExecJS==1.4.0
pylint==1.5.2
pymongo==3.4.0
pytest==3.0.7
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
Werkzeug==0.12.1
wrapt==1.8.0
xlrd==1.0.0
